LayoutTests/platform/ios-simulator/editing/pasteboard/copy-in-password-field-expected.txt

EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
This test verifies that copying is disabled for password fields by attempting to copy from a password field and paste into a textfield. If the test passes, you'll see a of 'PASS' message below, and the textfield will remain unmodified.

(The test is only meaningful in debug builds, since paste is disabled in release builds.)

  
PASS document.getElementById("text").value is "[not modified]"
